Follow-up commit for Homework 3 (05.10.2023)
