+++ src/main/java/ru/skypro/homework/HomeworkApplication.java
@@ -2,9 +2,11 @@ package ru.skypro.homework;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
 
 @SpringBootApplication
+//@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 public class HomeworkApplication {
   public static void main(String[] args) {
     SpringApplication.run(HomeworkApplication.class, args);
