+++ src/main/java/ru/skypro/homework/dto/Login.java
@@ -12,4 +12,20 @@ public class Login {
     private String username;
     @Size(min = 4, max = 32, message = "Проверьте количество символов.")
     private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
