+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -1,10 +1,10 @@
 package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.skypro.homework.model.User;
+import ru.skypro.homework.model.UserEntity;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
-    User getReferenceById(Long id);
+    UserEntity getReferenceById(Long id);
 
 }
