+++ src/main/java/ru/skypro/homework/service/mapping/AdMapping.java
@@ -1,11 +1,13 @@
 package ru.skypro.homework.service.mapping;
 
+import org.hibernate.annotations.Parameter;
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 import ru.skypro.homework.dto.Ad;
 import ru.skypro.homework.dto.CreateOrUpdateAd;
 import ru.skypro.homework.dto.ExtendedAd;
 
-@Mapper
+//@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface AdMapping {
 
     public Ad adEntityToAdDto(ru.skypro.homework.model.Ad ad);
