+++ src/main/java/ru/skypro/homework/service/mapping/CommentaryMapping.java
@@ -1,6 +1,7 @@
 package ru.skypro.homework.service.mapping;
 
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.dto.Comments;
 import ru.skypro.homework.dto.CreateOrUpdateComment;
@@ -10,7 +11,7 @@ import ru.skypro.homework.model.Commentary;
 import java.util.ArrayList;
 import java.util.List;
 
-@Mapper
+//@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface CommentaryMapping {
 
     public Comment commentEntityToCommentDto(Commentary commentary);
