+++ src/main/java/ru/skypro/homework/service/mapping/UserMapping.java
@@ -1,17 +1,18 @@
 package ru.skypro.homework.service.mapping;
 
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 import org.springframework.stereotype.Component;
 import ru.skypro.homework.dto.NewPassword;
 import ru.skypro.homework.dto.UpdateUser;
-import ru.skypro.homework.model.User;
+import ru.skypro.homework.model.UserEntity;
 
-@Mapper
+//@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface UserMapping {
 
-    public User newPasswordDtoToUser (NewPassword newPassword);
+    public UserEntity newPasswordDtoToUser (NewPassword newPassword);
 
-    public ru.skypro.homework.dto.User userEntityToUserDto (User userEntityInput);
+    public ru.skypro.homework.dto.User userEntityToUserDto (UserEntity userEntityInput);
 
-    public User updateUserDtoToUserEntity (UpdateUser updateUser) ;
+    public UserEntity updateUserDtoToUserEntity (UpdateUser updateUser) ;
 }
