+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/diplomaproject
 spring.datasource.username=postgres
 spring.datasource.password=123456
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
